Create nvidia-stats.sh
If you are wondering which process is causing your gpu to not go into S0ix state or suspend, this is for you. It tracks down the PIDs of the PIDs using your NVIDIA card.

Great for squeezing out more battery life from a laptop who refuses to suspend the NVIDIA card.